#39 modification de la position des boutons, changement de noms et suppression du champ de trop de studyQuestion
