+++ src/components/TsvLineForm.jsx
@@ -1,21 +1,18 @@
-import { Box, FormControl, TextField, Button, IconButton } from "@mui/material";
+import { Box, FormControl, TextField } from "@mui/material";
 import MarkdownField from "./MarkdownField";
 import { useState, useContext, useEffect } from "react";
 import {
     i18nContext as I18nContext,
     doI18n,
 } from "pithekos-lib";
-import DeleteIcon from '@mui/icons-material/Delete';
 import { v4 as uuidv4 } from 'uuid';
-import DeleteNote from "./DeleteNote";
 import ButtonsNavigation from "./ButtonsNavigation";
 import md5 from "md5";
 
-function TsvLineForm({ mode, currentRow, ingredient, setIngredient, saveFunction, currentRowN, setCurrentRowN, setSaveIngredientTsv, updateBcv }) {
+function TsvLineForm({ingredient, setIngredient, currentRowN, setCurrentRowN, setSaveIngredientTsv, updateBcv, mode, currentRow, saveFunction,  }) {
     const { i18nRef } = useContext(I18nContext);
-    const [rowData, setRowData] = useState(Array(7).fill("", 0, 7))
     const [cellValueChanged, setCellValueChanged] = useState(false);
-    const [openedModalDelete, setOpenedModalDelete] = useState(false);
+    const [rowData, setRowData] = useState(Array(7).fill("", 0, 7))
     const columnNames = ingredient[0] || [];
     const [md5Ingredient, setMd5Ingredient] = useState("");
 
@@ -39,10 +36,6 @@ function TsvLineForm({ mode, currentRow, ingredient, setIngredient, saveFunction
         console.log("Hash initial :", hash);
     }, []);
 
-    // Permet d'ouvrir la modal Delete
-    const handleOpenModalDelete = () => {
-        setOpenedModalDelete("delete");
-    };
 
     // Permet la modification d'une note
     const changeCell = (event, n) => {
@@ -105,81 +98,23 @@ function TsvLineForm({ mode, currentRow, ingredient, setIngredient, saveFunction
                 </FormControl>
 
             ))}
-            <Box sx={{ display: 'flex', gap: 2, justifyContent: "center" }}>
-                {mode === "add" && (<Button
-                    onClick={() => { saveFunction(currentRowN, rowData); setCellValueChanged(false) }}
-                    variant="contained"
-                    disabled={!cellValueChanged}
-                    sx={{
-                        mt: 2,
-                        "&.Mui-disabled": {
-                            background: "#eaeaea",
-                            color: "#424242"
-                        }
-                    }}
-                >
-                    {`${doI18n("pages:core-local-workspace:save_note", i18nRef.current)}`}
-                </Button>)}
+            <ButtonsNavigation
+                updateBcv={updateBcv}
+                rowData={rowData}
+                saveFunction={saveFunction}
+                handleCancel={handleCancel}
+                mode={mode}
+                setSaveIngredientTsv={setSaveIngredientTsv}
 
-                {mode === "edit" && (
-                    <>
-
-                        <Button
-                            onClick={() => {
-                                handleCancel();
-                                setCellValueChanged(false);
-                              
-                            }}
-                            variant="contained"
-                            disabled={!cellValueChanged}
-                            sx={{
-                                mt: 2,
-                                "&.Mui-disabled": {
-                                    background: "#eaeaea",
-                                    color: "#424242"
-                                }
-                            }}
-                        >
-                            {doI18n("pages:core-local-workspace:cancel", i18nRef.current)}
-                        </Button>
-
-                        <ButtonsNavigation
-                            updateBcv={updateBcv}
-                            rowData={rowData}
-
-                            setSaveIngredientTsv={setSaveIngredientTsv}
-
-                            ingredient={ingredient}
-                            setIngredient={setIngredient}
-
-                            currentRowN={currentRowN}
-                            setCurrentRowN={setCurrentRowN}
-                        />
-
-                        <IconButton
-                            onClick={() => handleOpenModalDelete()}
-                            sx={{
-                                mt: 2
-                            }}
-                        >
-                            <DeleteIcon size="large" color="primary" />
-                        </IconButton>
-                    </>
-                )}
-
-            </Box>
-            <DeleteNote
-                mode="delete"
-                open={openedModalDelete === "delete"}
-                closeModal={() => setOpenedModalDelete(null)}
                 ingredient={ingredient}
                 setIngredient={setIngredient}
-                rowData={rowData}
-                setSaveIngredientTsv={setSaveIngredientTsv}
-                currentRowN={currentRowN}
-               
-            />
 
+                currentRowN={currentRowN}
+                setCurrentRowN={setCurrentRowN}
+
+                cellValueChanged={cellValueChanged}
+                setCellValueChanged={setCellValueChanged}
+            />
         </Box>
 
     )
